// api/api.ts
export interface Shipyard {
    /**
     * The symbol of the shipyard. The symbol is the same as the waypoint where the shipyard is located.
     * @type {string}
     * @memberof Shipyard
     */
    'symbol': string;
    /**
     * The list of ship types available for purchase at this shipyard.
     * @type {Array<ShipyardShipTypesInner>}
     * @memberof Shipyard
     */
    'shipTypes': Array<ShipyardShipTypesInner>;
    /**
     * The list of recent transactions at this shipyard.
     * @type {Array<ShipyardTransaction>}
     * @memberof Shipyard
     */
    'transactions'?: Array<ShipyardTransaction>;
    /**
     * The ships that are currently available for purchase at the shipyard.
     * @type {Array<ShipyardShip>}
     * @memberof Shipyard
     */
    'ships'?: Array<ShipyardShip>;
    /**
     * The fee to modify a ship at this shipyard. This includes installing or removing modules and mounts on a ship. In the case of mounts, the fee is a flat rate per mount. In the case of modules, the fee is per slot the module occupies.
     * @type {number}
     * @memberof Shipyard
     */
    'modificationsFee': number;
}
/**
 * 
 * @export
 * @interface ShipyardShip
 */
export interface ShipyardShip {
    /**
     * 
     * @type {ShipType}
     * @memberof ShipyardShip
     */
    'type': ShipType;
    /**
     * 
     * @type {string}
     * @memberof ShipyardShip
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof ShipyardShip
     */
    'description': string;
    /**
     * 
     * @type {SupplyLevel}
     * @memberof ShipyardShip
     */
    'supply': SupplyLevel;
    /**
     * 
     * @type {ActivityLevel}
     * @memberof ShipyardShip
     */
    'activity'?: ActivityLevel;
    /**
     * 
     * @type {number}
     * @memberof ShipyardShip
     */
    'purchasePrice': number;
    /**
     * 
     * @type {ShipFrame}
     * @memberof ShipyardShip
     */
    'frame': ShipFrame;
    /**
     * 
     * @type {ShipReactor}
     * @memberof ShipyardShip
     */
    'reactor': ShipReactor;
    /**
     * 
     * @type {ShipEngine}
     * @memberof ShipyardShip
     */
    'engine': ShipEngine;
    /**
     * 
     * @type {Array<ShipModule>}
     * @memberof ShipyardShip
     */
    'modules': Array<ShipModule>;
    /**
     * 
     * @type {Array<ShipMount>}
     * @memberof ShipyardShip
     */
    'mounts': Array<ShipMount>;
    /**
     * 
     * @type {ShipyardShipCrew}
     * @memberof ShipyardShip
     */
    'crew': ShipyardShipCrew;
}


/**
 * 
 * @export
 * @interface ShipyardShipCrew
 */
export interface ShipyardShipCrew {
    /**
     * 
     * @type {number}
     * @memberof ShipyardShipCrew
     */
    'required': number;
    /**
     * 
     * @type {number}
     * @memberof ShipyardShipCrew
     */
    'capacity': number;
}
/**
 * 
 * @export
 * @interface ShipyardShipTypesInner
 */
export interface ShipyardShipTypesInner {
    /**
     * 
     * @type {ShipType}
     * @memberof ShipyardShipTypesInner
     */
    'type': ShipType;
}


/**
 * Results of a transaction with a shipyard.
 * @export
 * @interface ShipyardTransaction
 */
export interface ShipyardTransaction {
    /**
     * The symbol of the waypoint.
     * @type {string}
     * @memberof ShipyardTransaction
     */
    'waypointSymbol': string;
    /**
     * The symbol of the ship that was the subject of the transaction.
     * @type {string}
     * @memberof ShipyardTransaction
     * @deprecated
     */
    'shipSymbol': string;
    /**
     * The symbol of the ship that was the subject of the transaction.
     * @type {string}
     * @memberof ShipyardTransaction
     */
    'shipType': string;
    /**
     * The price of the transaction.
     * @type {number}
     * @memberof ShipyardTransaction
     */
    'price': number;
    /**
     * The symbol of the agent that made the transaction.
     * @type {string}
     * @memberof ShipyardTransaction
     */
    'agentSymbol': string;
    /**
     * The timestamp of the transaction.
     * @type {string}
     * @memberof ShipyardTransaction
     */
    'timestamp': string;
}
// ...
export type TradeSymbol = typeof TradeSymbol[keyof typeof TradeSymbol];
// ...
export const WaypointTraitSymbol = {
    Uncharted: 'UNCHARTED',
    UnderConstruction: 'UNDER_CONSTRUCTION',
    Marketplace: 'MARKETPLACE',
    Shipyard: 'SHIPYARD',
    Outpost: 'OUTPOST',
    ScatteredSettlements: 'SCATTERED_SETTLEMENTS',
    SprawlingCities: 'SPRAWLING_CITIES',
    MegaStructures: 'MEGA_STRUCTURES',
    PirateBase: 'PIRATE_BASE',
    Overcrowded: 'OVERCROWDED',
    HighTech: 'HIGH_TECH',
    Corrupt: 'CORRUPT',
    Bureaucratic: 'BUREAUCRATIC',
    TradingHub: 'TRADING_HUB',
    Industrial: 'INDUSTRIAL',
    BlackMarket: 'BLACK_MARKET',
    ResearchFacility: 'RESEARCH_FACILITY',
    MilitaryBase: 'MILITARY_BASE',
    SurveillanceOutpost: 'SURVEILLANCE_OUTPOST',
    ExplorationOutpost: 'EXPLORATION_OUTPOST',
    MineralDeposits: 'MINERAL_DEPOSITS',
    CommonMetalDeposits: 'COMMON_METAL_DEPOSITS',
    PreciousMetalDeposits: 'PRECIOUS_METAL_DEPOSITS',
    RareMetalDeposits: 'RARE_METAL_DEPOSITS',
    MethanePools: 'METHANE_POOLS',
    IceCrystals: 'ICE_CRYSTALS',
    ExplosiveGases: 'EXPLOSIVE_GASES',
    StrongMagnetosphere: 'STRONG_MAGNETOSPHERE',
    VibrantAuroras: 'VIBRANT_AURORAS',
    SaltFlats: 'SALT_FLATS',
    Canyons: 'CANYONS',
    PerpetualDaylight: 'PERPETUAL_DAYLIGHT',
    PerpetualOvercast: 'PERPETUAL_OVERCAST',
    DrySeabeds: 'DRY_SEABEDS',
    MagmaSeas: 'MAGMA_SEAS',
    Supervolcanoes: 'SUPERVOLCANOES',
    AshClouds: 'ASH_CLOUDS',
    VastRuins: 'VAST_RUINS',
    MutatedFlora: 'MUTATED_FLORA',
    Terraformed: 'TERRAFORMED',
    ExtremeTemperatures: 'EXTREME_TEMPERATURES',
    ExtremePressure: 'EXTREME_PRESSURE',
    DiverseLife: 'DIVERSE_LIFE',
    ScarceLife: 'SCARCE_LIFE',
    Fossils: 'FOSSILS',
    WeakGravity: 'WEAK_GRAVITY',
    StrongGravity: 'STRONG_GRAVITY',
    CrushingGravity: 'CRUSHING_GRAVITY',
    ToxicAtmosphere: 'TOXIC_ATMOSPHERE',
    CorrosiveAtmosphere: 'CORROSIVE_ATMOSPHERE',
    BreathableAtmosphere: 'BREATHABLE_ATMOSPHERE',
    ThinAtmosphere: 'THIN_ATMOSPHERE',
    Jovian: 'JOVIAN',
    Rocky: 'ROCKY',
    Volcanic: 'VOLCANIC',
    Frozen: 'FROZEN',
    Swamp: 'SWAMP',
    Barren: 'BARREN',
    Temperate: 'TEMPERATE',
    Jungle: 'JUNGLE',
    Ocean: 'OCEAN',
    Radioactive: 'RADIOACTIVE',
    MicroGravityAnomalies: 'MICRO_GRAVITY_ANOMALIES',
    DebrisCluster: 'DEBRIS_CLUSTER',
    DeepCraters: 'DEEP_CRATERS',
    ShallowCraters: 'SHALLOW_CRATERS',
    UnstableComposition: 'UNSTABLE_COMPOSITION',
    HollowedInterior: 'HOLLOWED_INTERIOR',
    Stripped: 'STRIPPED'
} as const;

export type WaypointTraitSymbol = typeof WaypointTraitSymbol[keyof typeof WaypointTraitSymbol];